16 medium C++

+++ C++/Medium/16.cpp
@@ -10,6 +10,6 @@ void print(float a)
 }
 int main()
 {
-    print(5.0);
+    print(static_cast<float>(5.0)); // any decimal numbers are regarded as doubles in C++ not float so we need to epllicity state it's a float so it can find proper function
     return 0;
 }
